Refactor userid router
routes/user/userid 아래의 라우터들을 user 와 같은 레벨로 옮김
