+++ src/routes/user/category/index.ts
@@ -1,5 +1,5 @@
 import { Router } from 'express';
-import helloWorld from '../../../hello-world';
+import helloWorld from '../../hello-world';
 
 const router = Router();
 // 사용자의 관심 카테고리 목록 읽어오기
