+++ src/routes/user/index.ts
@@ -1,21 +1,29 @@
 import { Request, Response, NextFunction, Router } from 'express';
-import userIdRouter from './userid';
 import profileRouter from './profile';
 import loginRouter from './login';
+import roleRouter from './role';
+import notificationRouter from './notification';
+import categoryRouter from './category';
 import { saveUser } from '../../services/User';
+import helloWorld from '../hello-world';
 
 const router = Router();
 router.post('/', saveUser);
+router.patch('/', helloWorld);
+router.delete('/', helloWorld);
 router.use(
-  '/:id',
+  '/:id/role',
   (req: Request, res: Response, next: NextFunction) => {
     req.user = {
       id: req.params.id,
     };
     next();
   },
-  userIdRouter
+  roleRouter
 );
+router.use('/:id/notification', notificationRouter);
+router.use('/:id/category', categoryRouter);
+
 router.use('/profile', profileRouter);
 router.use('/login', loginRouter);
