+++ src/routes/user/notification/index.ts
@@ -1,5 +1,5 @@
 import { Router } from 'express';
-import helloWorld from '../../../hello-world';
+import helloWorld from '../../hello-world';
 
 const router = Router();
 router.get('/', helloWorld);
