+++ src/routes/user/role/index.ts
@@ -1,5 +1,5 @@
 import { Router } from 'express';
-import { changeUserRole } from '../../../../services/User';
+import { changeUserRole } from '../../../services/User';
 
 const router = Router();
 router.patch('/', changeUserRole);
